pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data_ev"
version = "1.0.0"
description = "Electric Vehicle Data Processing Pipeline"
readme = "readme.md"
requires-python = ">=3.11"
dependencies = [
    # Core dependencies used across all modules
    "python-dotenv==1.0.0",
    "pandas==2.2.2", # Adds 150MB to images, might be worth to install only in needed modules
    "msgspec==0.19.0",
    "schedule==1.2.2",
    "rich==13.5.2",
    "requests==2.30.0",
    "redis>=6.0.0",
    "asyncio-mqtt",
    "aiohttp==3.11.11",
    "aioboto3>=13.4.0",
    "boto3>=1.36",
]

[project.optional-dependencies]
# Activation module dependencies - handles vehicle API communication
activation = [
    "gspread",
    "oauth2client",
    "psycopg2-binary>=2.9.9",
    "asyncpg~=0.30",
    "uuid6==2022.10.25",
    "sqlalchemy>=2.0.0",
]

# Transform module dependencies - handles ETL pipeline with Spark
transform = [
    "pyspark==3.5.5",
    "pyarrow==19.0.0",
    "apscheduler==3.10.4",
    "psycopg2-binary>=2.9.9",
    "asyncpg~=0.30",
    "minio==7.1.13",
    "pydantic>=1.10.13",
    "pydantic-settings~=2.11",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0.0",
]

# EDA module dependencies - for exploratory data analysis notebooks
eda = [
    "plotly==5.19.0",
    "matplotlib==3.7.2",
    "scipy==1.10.1",
    "scikit-learn==1.5.1",
    "statsmodels==0.14.4",
    "pyarrow==19.0.0",
    "nbformat==5.10.4",
    "nbstripout==0.8.1",
    "tqdm==4.66.5",
]

# Load module dependencies - for loading and processing trendlines
load = [
    "rapidfuzz==3.13.0",
    "gspread",
    "oauth2client",
    "psycopg2-binary>=2.9.9",
    "asyncpg~=0.30",
    "sqlmodel==0.0.24",
]

# Web module dependencies - FastAPI web application
web = [
    "fastapi~=0.118",
    "uvicorn",
    "pydantic>=1.10.13",
    "pydantic-settings~=2.11",
    "python-multipart>=0.0.18",
    "orjson==3.8.5",
]

# Results/visualization module dependencies
results = [
    "streamlit==1.47.0",
    "plotly==5.19.0",
    "psycopg2-binary>=2.9.9",
    "gspread",
    "oauth2client",
]

# Ingestion module dependencies - for data ingestion from various sources
ingestion = [
    "rapidfuzz==3.13.0",
    "aiokafka==0.10.0",
    "memory-profiler==0.61.0",
]

# Scraping module dependencies
scraping = [
    "beautifulsoup4==4.13.4",
    "selenium==4.33.0",
    "PyPDF2==3.0.1",
    "lxml==6.0.1",
]

external_api = [
    "bcrypt==3.2.2",
    "fastapi~=0.118",
    "fastapi-pagination~=0.14",
    "uvicorn[standard]~=0.23.2",
    "python-jose[cryptography]~=3.3.0",
    "python-multipart~=0.0.6",
    "passlib[bcrypt]==1.7.4",
    "pydantic~=2.11",
    "pydantic[email]~=2.11",
    "pydantic-settings~=2.11",
    "sqlalchemy~=2.0.20",
    "sqlmodel~=0.0.8",
    "asyncpg~=0.30",
    "psycopg2-binary~=2.9.7",
    "redis~=6.0.0",
    "python-dotenv~=1.0.0",
    "structlog~=23.1.0", # Logging structuré"
    "oauth2client~=4.1.3", # Pour l'authentification avec Google API "
    "boto3~=1.36", # S3 integration",
    "pandas~=2.2", # Data analysis",
    "joblib==1.3.2", # Joblib for machine learning"
    "asteval>=1.0.6",
    "gspread>=6.1.4",
]

[dependency-groups]
# Full development environment with all dependencies
dev = [
    "data_ev[activation,transform,eda,load,web,results,ingestion,external_api]",
    "pytest",
    "ruff",
    "nbstripout",
    "fakeredis>=2.31.3",
    "py-spy>=0.4.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 88
extend-exclude = ["*.ipynb", "*.csv"]

[tool.ruff.lint]
select = [
    "E", # pycodestyle (error)
    "F", # pyflakes
    "B", # bugbear
    "B9",
    "C4", # flake8-comprehensions
    "SIM", # flake8-simplify
    "I", # isort
    "UP", # pyupgrade
    "PIE", # flake8-pie
    "PGH", # pygrep-hooks
    "PYI", # flake8-pyi
    "RUF",
]

ignore = [
    # only relevant if you run a script with `python -0`,
    # which seems unlikely for any of the scripts in this repo
    "B011",
    # Leave it to the formatter to split long lines and
    # the judgement of all of us.
    "E501"
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["core", "activation", "transform", "eda", "load", "web", "results", "external_api"]

